requirements.txt
Flask
Flask-WTF
python-dotenv